+++ doc/source/changelog.md
@@ -14,6 +14,7 @@ Primarily developped with Godot 4.3.
 
 - Added project setting `voxel/ownership_checks` to turn off sanity checks done by certain virtual functions that pass an object (such as `_generate_block`). Relevant for C#, where the garbage collection model prevents such checks from working properly.
 - `VoxelBuffer`: Added several functions to do arithmetic operations on all voxels
+- `VoxelInstanceGenerator`: allow to set density beyond 1, up to 10, by typing it in the field
 - `VoxelMesherBlocky`:
     - Can be used with `VoxelLodTerrain`. Basic support: meshes scale with LOD and LOD>1 chunks have extra geometry to reduce cracks between LODs
     - Added experimental "shadow occluders": generates quads on chunk sides if they are covered by opaque voxels, to force shadows to project in caves when  there is no surface for DirectionalLight to project from (see #622).